allow filesystem source
chooses a source based on the identifier: if it starts with "Masters/", then use the FilesystemSource. Otherwise, use HttpSource as before to retreive from fedora.

(the slash must be url-escaped as %2F in the request, but it is already unescaped by the time the delegate script is invoked)

Also adds the method `authorize`, which replaces `authorized?` and `redirect` in Cantaloupe 4.1 (https://cantaloupe-project.github.io/manual/4.1/delegate-script.html#MigratingFrom4To41)